requirements.txt
chromadb==0.4.24
pypdf==4.3.1
python-dotenv==1.0.1
requests==2.32.3
nltk==3.9.1
fastapi==0.115.5
uvicorn==0.30.6
jinja2==3.1.4
pytest==7.4.4
